Scaffold J3DMtxBuffer.cpp

+++ include/JSystem/J3D/J3DJointTree.h
@@ -31,27 +31,27 @@ struct J3DJointTree {
 	// unused/inlined:
 	void clear();
 
-	J3DJoint* getJointNodePointer(u16 idx) const { return mJoints[idx]; }
-	u16 getWEvlpMtxNum() const { return mEnvelopeCnt; }
+	J3DJoint* getJointNodePointer(u32 idx) const { return mJoints[idx]; }
+	u32 getWEvlpMtxNum() const { return mEnvelopeCnt; }
 	u8 getWEvlpMixMtxNum(u16 idx) const { return mEnvelopeMixCnt[idx]; }
 	u16* getWEvlpMixIndex() const { return mEnvelopeMixIdx; }
 	f32* getWEvlpMixWeight() const { return mEnvelopeMixWeight; }
 	u16* getWEvlpImportantMtxIndex() const { return mEnvelopeImptIdx; }
 	u32 getDrawFullWgtMtxNum() const { return mMtxData.mDrawMtxCount; }
-	u16 getJointNum() const { return mJointCnt; }
+	u32 getJointNum() const { return mJointCnt; }
 	u16 getDrawMtxNum() const { return mMtxData.mCount; }
 	u8 getDrawMtxFlag(u16 idx) const { return mMtxData.mDrawMtxFlag[idx]; }
-	u16 getDrawMtxIndex(u16 idx) const { return mMtxData.mDrawMtxIdx[idx]; }
+	u32 getDrawMtxIndex(u16 idx) const { return mMtxData.mDrawMtxIdx[idx]; }
 	JUTNameTab* getJointName() const { return mNametab; }
 	J3DJoint* getRootNode() { return mRootNode; }
 	J3DMtxCalc* getBasicMtxCalc() const { return mTransformCalc; }
-	Mtx& getInvJointMtx(s32 idx) const { return mInvJointMtx[idx]; }
+	Mtx& getInvJointMtx(int idx) const { return mInvJointMtx[idx]; }
 	u32 getModelDataType() const { return mModelDataType; }
 	const J3DModelHierarchy* getHierarchy() const { return mHierarchy; }
 
 	// VTBL _00
 	J3DModelHierarchy* mHierarchy; // _04
-	u32 mFlags;                    // _08 previously s8* TODO: rename
+	u32 mFlags;                    // _08
 	u32 mModelDataType;            // _0C
 	J3DJoint* mRootNode;           // _10
 	J3DMtxCalc* mTransformCalc;    // _14
